backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
python-multipart
